refactor: rename models/meshmodel to models/registry, add compat shim for schemas@v1.3.26
