+++ models/registry/entity/error.go
@@ -1,0 +1,15 @@
+package entity
+
+import (
+	"fmt"
+
+	"github.com/meshery/meshkit/errors"
+)
+
+const (
+	ErrUpdateEntityStatusCode = "meshkit-11243"
+)
+
+func ErrUpdateEntityStatus(err error, entity string, status EntityStatus) error {
+	return errors.New(ErrUpdateEntityStatusCode, errors.Alert, []string{fmt.Sprintf("unable to update %s to %s", entity, status)}, []string{err.Error()}, []string{}, []string{})
+}
